[Issue 6] merged pravega writer and exactly once writer code (#30)
[issue-6] Support EXACTLY_ONCE writer mode in `FlinkPravegaWriter`, remove `FlinkPravegaExactlyOnceWriter`.
